Move Action classes to their own package.
Actions are becoming quite central to MWM and the ActionsApp class was becoming a bit crowded...
