+++ src/org/metawatch/manager/actions/ResettableAction.java
@@ -1,0 +1,7 @@
+package org.metawatch.manager.actions;
+
+import android.content.Context;
+
+public interface ResettableAction extends Action {
+	public int performReset(Context context);
+}
