+++ src/org/metawatch/manager/actions/TimestampAction.java
@@ -1,0 +1,5 @@
+package org.metawatch.manager.actions;
+
+public interface TimestampAction extends Action {
+	public long getTimestamp();
+}
